Add embedding-based ads serving functionality (#17936)
* Add string initialization to VectorData

* Add embedding table and parsing

* Add unittests

* Change BuildUserModel to async

* Add database schema 27

* Add exclusion rule for ads with no embeddings in v3

* Refactor exclusion rule from per hour to creative instance

* Move notification specific exclusion rules

* Use IsApproximatelyEqual when comparing floats

---------

+++ components/brave_ads/core/internal/ml/pipeline/embedding_pipeline_value_util.cc
@@ -58,7 +58,7 @@ absl::optional<EmbeddingPipelineInfo> EmbeddingPipelineFromValue(
     std::vector<float> embedding;
     embedding.reserve(list->size());
     for (const base::Value& dimension_value : *list) {
-      embedding.push_back(dimension_value.GetDouble());
+      embedding.push_back(static_cast<float>(dimension_value.GetDouble()));
     }
     embedding_pipeline.embeddings[embedding_key] =
         VectorData(std::move(embedding));
